backuptool_ab: Use mv instead of cp for restore step
Using cp instead of mv during the restore step resulted in double the
amount of free space and inodes being used.
For example, a nano arm64 opengapps install uses around 500M. But during
a backup/restore process when the Updater is used, it temporarily uses
1G on the other slot.  That causes failed updates on devices that
don't have that much free space on the system partition.

Change-Id: I57ba2e0f52d328d4ff91d47a3db45f47bb8402af
Signed-off-by: AshutoshSundresh <ashutoshsundresh@gmail.com>