overlay: SystemUI: Use Monet colors for power menu
It's odd that the power menu doesn't use Monet colors, unlike most of
System UI. This makes it feel more integrated with the rest of the
system.

 overlay: SystemUI: Follow light/dark theme in power menu

Now that we've modified the power menu to refresh on UI mode changes
and use Monet colors, make it follow the system light/dark theme for
better integration in light mode.

Requires frameworks/base commit:
    SystemUI: Always refresh global actions dialog on UI mode change

 overlay: SystemUI: Remove power menu shadow

This fits better with Material You, especially when the default shadow
gets clipped at the edges.

 overlay: SystemUI: Use default surface colors for power menu buttons

The Settings search bar uses surfaceVariant in dark mode and
surfaceHighlight in light mode, but the default surface colors work
better here.